+++ packages/pluggableWidgets/badge-button-web/CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   We changed colors in the structure mode preview for dark and light modes.
+
 -   We updated page explorer's caption to display label.
+
 -   We updated the light and dark icons and tiles for the widget.
 
 ## [3.1.0] - 2021-12-23